+++ packages/pintora-core/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change Log
 
+## 0.3.0-alpha.2
+
+### Patch Changes
+
+- bc16cae: [gantt] add support for axisInterval
+- 46a1a9f: [gantt] add support for axisInterval
+
+## 0.3.0-alpha.1
+
+### Patch Changes
+
+- 235cc0e: add @config and @param to gantt diagram; update tmLanguage;
+
+## 0.3.0-alpha.0
+
+### Minor Changes
+
+- bd214c7: Add gantt diagram
+
 ## 0.2.0
 
 ### Minor Changes