+++ website/CHANGELOG.md
@@ -1,5 +1,37 @@
 # Change Log
 
+## 0.3.0-alpha.2
+
+### Patch Changes
+
+- bc16cae: [gantt] add support for axisInterval
+- 46a1a9f: [gantt] add support for axisInterval
+- Updated dependencies [bc16cae]
+- Updated dependencies [46a1a9f]
+  - @pintora/standalone@0.3.0-alpha.2
+  - @pintora/test-shared@0.2.0-alpha.3
+
+## 0.3.0-alpha.1
+
+### Patch Changes
+
+- 235cc0e: add @config and @param to gantt diagram; update tmLanguage;
+- Updated dependencies [235cc0e]
+  - @pintora/standalone@0.3.0-alpha.1
+  - @pintora/test-shared@0.2.0-alpha.2
+
+## 0.3.0-alpha.0
+
+### Minor Changes
+
+- bd214c7: Add gantt diagram
+
+### Patch Changes
+
+- Updated dependencies [bd214c7]
+  - @pintora/test-shared@0.2.0-alpha.1
+  - @pintora/standalone@0.3.0-alpha.0
+
 ## 0.2.0
 
 ### Minor Changes